Splitted into the dif components, fixed hashRouting issues
