+++ src/components/PeopleTable/PersonTable.tsx
@@ -1,0 +1,45 @@
+import React from 'react';
+import { Person } from '../../types';
+import { PersonLink } from '../PersonLink/PersonLink';
+
+type Props = {
+  people: Person[];
+};
+
+export const PersonTable: React.FC<Props> = ({ people }) => {
+  const findPersonBySlug = (name: string | null) => {
+    if (!name) {
+      return null;
+    }
+
+    return people.find(person => person.name === name);
+  };
+
+  return (
+    <table
+      data-cy="peopleTable"
+      className="table is-striped is-hoverable is-narrow is-fullwidth"
+    >
+      <thead>
+        <tr>
+          <th>Name</th>
+          <th>Sex</th>
+          <th>Born</th>
+          <th>Died</th>
+          <th>Mother</th>
+          <th>Father</th>
+        </tr>
+      </thead>
+
+      <tbody>
+        {people.map(person => (
+          <PersonLink
+            key={person.slug}
+            person={person}
+            findPersonBySlug={findPersonBySlug}
+          />
+        ))}
+      </tbody>
+    </table>
+  );
+};
