+++ src/pages/PeoplePage/PeoplePage.tsx
@@ -1,15 +1,13 @@
 import { useEffect, useState } from 'react';
-import { useParams, useNavigate } from 'react-router-dom';
 import { Loader } from '../../components/Loader/Loader';
 import { getPeople } from '../../api';
 import { Person } from '../../types/Person';
+import { PersonTable } from '../../components/PeopleTable/PersonTable';
 
 export const PeoplePage = () => {
   const [people, setPeople] = useState<Person[]>([]);
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState(false);
-  const { slug } = useParams();
-  const navigate = useNavigate();
 
   useEffect(() => {
     setIsLoading(true);
@@ -20,18 +18,6 @@ export const PeoplePage = () => {
       .catch(() => setError(true))
       .finally(() => setIsLoading(false));
   }, []);
-
-  const findPersonBySlug = (name: string | null) => {
-    if (!name) {
-      return null;
-    }
-
-    return people.find(person => person.name === name);
-  };
-
-  const handlePersonClick = (personSlug: string) => {
-    navigate(`/people/${personSlug}`, { replace: true });
-  };
 
   return (
     <>
@@ -52,100 +38,7 @@ export const PeoplePage = () => {
           )}
 
           {!isLoading && !error && people.length > 0 && (
-            <table
-              data-cy="peopleTable"
-              className="table is-striped is-hoverable is-narrow is-fullwidth"
-            >
-              <thead>
-                <tr>
-                  <th>Name</th>
-                  <th>Sex</th>
-                  <th>Born</th>
-                  <th>Died</th>
-                  <th>Mother</th>
-                  <th>Father</th>
-                </tr>
-              </thead>
-
-              <tbody>
-                {people.map(person => {
-                  const mother = findPersonBySlug(person.motherName);
-                  const father = findPersonBySlug(person.fatherName);
-
-                  return (
-                    <tr
-                      key={person.slug}
-                      data-cy="person"
-                      className={
-                        person.slug === slug ? 'has-background-warning' : ''
-                      }
-                    >
-                      <td>
-                        <a
-                          className={
-                            person.sex === 'f' ? 'has-text-danger' : ''
-                          }
-                          href={`#/people/${person.slug}`}
-                          onClick={e => {
-                            e.preventDefault();
-                            handlePersonClick(person.slug);
-                          }}
-                        >
-                          {person.name}
-                        </a>
-                      </td>
-                      <td>{person.sex}</td>
-                      <td>{person.born}</td>
-                      <td>{person.died}</td>
-                      <td>
-                        {person.motherName ? (
-                          mother ? (
-                            <a
-                              className={
-                                mother.sex === 'f' ? 'has-text-danger' : ''
-                              }
-                              href={`#/people/${mother.slug}`}
-                              onClick={e => {
-                                e.preventDefault();
-                                handlePersonClick(mother.slug);
-                              }}
-                            >
-                              {person.motherName}
-                            </a>
-                          ) : (
-                            person.motherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                      <td>
-                        {person.fatherName ? (
-                          father ? (
-                            <a
-                              className={
-                                father.sex === 'f' ? 'has-text-danger' : ''
-                              }
-                              href={`#/people/${father.slug}`}
-                              onClick={e => {
-                                e.preventDefault();
-                                handlePersonClick(father.slug);
-                              }}
-                            >
-                              {person.fatherName}
-                            </a>
-                          ) : (
-                            person.fatherName
-                          )
-                        ) : (
-                          '-'
-                        )}
-                      </td>
-                    </tr>
-                  );
-                })}
-              </tbody>
-            </table>
+            <PersonTable people={people} />
           )}
         </div>
       </div>
